+++ api/pkg/webhooks/client.go
@@ -168,3 +168,32 @@ func (g *simpleWebhookClient) GetUseDataFrom() string {
 func (g *simpleWebhookClient) GetName() string {
 	return g.Name
 }
+
+func validateWebhookConfig(webhookConfig *WebhookConfig) error {
+	if webhookConfig.Name == "" {
+		return fmt.Errorf("missing webhook name")
+	}
+	if webhookConfig.URL == "" {
+		return fmt.Errorf("missing webhook URL")
+	}
+	if webhookConfig.Method == "" {
+		webhookConfig.Method = http.MethodPost // Default to POST, TODO: decide if GET is allowed
+	}
+	if webhookConfig.AuthEnabled && webhookConfig.AuthToken == "" {
+		return fmt.Errorf("missing webhook auth token")
+	}
+	if webhookConfig.OnError == "" {
+		webhookConfig.OnError = onErrorAbort
+	}
+	if webhookConfig.OnError != onErrorAbort && webhookConfig.OnError != onErrorIgnore {
+		return fmt.Errorf("onError must be either 'abort' or 'ignore'")
+	}
+	if webhookConfig.NumRetries < -1 {
+		return fmt.Errorf("numRetries must be a positive integer or -1")
+	}
+	if webhookConfig.Timeout == nil {
+		def := 10
+		webhookConfig.Timeout = &def
+	}
+	return nil
+}
